Changed the Opinfo structure to include Callable functions for `FusionDefinition` patterns (#570)
