+++ manual_ci.sh
@@ -18,6 +18,7 @@ run_test() {
 cd "$(dirname "${BASH_SOURCE[0]}")"
 
 run_test './bin/nvfuser_tests'
+run_test 'pytest python_tests/pytest_ops.py'
 run_test 'python python_tests/test_dynamo.py'
 run_test 'python python_tests/test_python_frontend.py'
 run_test 'python python_tests/test_schedule_ops.py'
