+++ python_tests/pytest_opinfos.py
@@ -6,6 +6,7 @@
 import torch
 import jax
 from pytest_core import OpInfo, ReferenceType, Domain
+from pytest_fusion_definitions import api_test_fd_fn, tensor_input_fd_fn
 from pytest_input_generators import (
     broadcast_error_generator,
     broadcast_in_dim_generator,
@@ -52,7 +53,8 @@ define_tensor_opinfo = OpInfo(
     "define_tensor",
     sample_input_generator=define_tensor_generator,
     error_input_generator=define_tensor_error_generator,
-    is_fusion_input_op=True,
+    fd_correctness_fn=tensor_input_fd_fn,
+    fd_error_input_fn=tensor_input_fd_fn,
 )
 fusion_input_ops.append(define_tensor_opinfo)
 
@@ -64,7 +66,7 @@ define_vector_constant_opinfo = OpInfo(
     "define_vector_constant",
     sample_input_generator=None,
     error_input_generator=define_vector_constant_error_generator,
-    is_fusion_input_op=True,
+    fd_error_input_fn=api_test_fd_fn,
 )
 fusion_input_ops.append(define_vector_constant_opinfo)
 
@@ -73,7 +75,7 @@ define_vector_input_opinfo = OpInfo(
     "define_vector_input",
     sample_input_generator=None,
     error_input_generator=define_vector_input_error_generator,
-    is_fusion_input_op=True,
+    fd_error_input_fn=api_test_fd_fn,
 )
 fusion_input_ops.append(define_vector_input_opinfo)
